mux: support SYNC/PSYNC and diskless replication for mux virtual clients
- Remove SYNC/PSYNC restriction for mux virtual clients in syncCommand
- Adapt diskless replication (rdbSaveToSlavesSockets) to use client* instead of connection* for rdb_pipe_conns, enabling transparent mux support
- Use muxConnWrite/muxConnSetWriteHandler in diskless RDB pipe handlers
- Move rdb_pipe_conns cleanup in freeClient before conn NULL check
- Fix format warning in mux.c: use %llu for uint64_t stream_id

+++ src/mux.c
@@ -399,8 +399,8 @@ static int muxProcessFrame(muxConnection *mux, uint8_t flags, uint64_t stream_id
 
     case MUX_FRAME_ERROR: {
         /* Stream-level error from client - log and close stream */
-        serverLog(LL_WARNING, "MUX: received error frame on stream %u from client %llu",
-                  stream_id, (unsigned long long)mux->owner_client->id);
+        serverLog(LL_WARNING, "MUX: received error frame on stream %llu from client %llu",
+                  (unsigned long long)stream_id, (unsigned long long)mux->owner_client->id);
         muxStream *ms = muxStreamLookup(mux, stream_id);
         if (ms) {
             ms->state = MUX_STREAM_CLOSED;
